requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
pypdf
groq